+++ src/components/expanded-state/unique-token/NFTBriefTokenInfoRow.tsx
@@ -14,6 +14,8 @@ import { ethereumUtils } from '@/utils';
 import { useNFTListing } from '@/resources/nfts';
 import { UniqueAsset } from '@/entities';
 import { fetchReservoirNFTFloorPrice } from '@/resources/nfts/utils';
+import { handleReviewPromptAction } from '@/utils/reviewAlert';
+import { ReviewPromptAction } from '@/storage/schema';
 
 const NONE = 'None';
 
@@ -30,6 +32,7 @@ export default function NFTBriefTokenInfoRow({
 }: {
   asset: UniqueAsset;
 }) {
+  const [hasDispatchedAction, setHasDispatchedAction] = useState(false);
   const { colors } = useTheme();
 
   const { navigate } = useNavigation();
@@ -73,10 +76,18 @@ export default function NFTBriefTokenInfoRow({
   );
 
   const [showFloorInEth, setShowFloorInEth] = useState(true);
-  const toggleFloorDisplayCurrency = useCallback(
-    () => setShowFloorInEth(!showFloorInEth),
-    [showFloorInEth, setShowFloorInEth]
-  );
+  const toggleFloorDisplayCurrency = useCallback(() => {
+    if (!hasDispatchedAction) {
+      handleReviewPromptAction(ReviewPromptAction.NftFloorPriceVisit);
+      setHasDispatchedAction(true);
+    }
+    setShowFloorInEth(!showFloorInEth);
+  }, [
+    showFloorInEth,
+    setShowFloorInEth,
+    hasDispatchedAction,
+    setHasDispatchedAction,
+  ]);
 
   const handlePressCollectionFloor = useCallback(() => {
     navigate(Routes.EXPLAIN_SHEET, {