+++ src/components/remote-promo-sheet/check-fns/isAfterCampaignLaunch.ts
@@ -1,0 +1,5 @@
+import { PromoSheet } from '@/graphql/__generated__/arc';
+
+export const isAfterCampaignLaunch = ({ launchDate }: PromoSheet): boolean => {
+  return new Date() > new Date(launchDate);
+};