+++ src/hooks/useSwapAdjustedAmounts.ts
@@ -7,6 +7,8 @@ import {
   WETH_ADDRESS,
   WMATIC_POLYGON_ADDRESS,
   WBNB_BSC_ADDRESS,
+  WETH_ARBITRUM_ADDRESS,
+  WETH_ZORA_ADDRESS,
 } from '@/references';
 import { fromWei, updatePrecisionToDisplay } from '@/helpers/utilities';
 import { ethereumUtils } from '@/utils';
@@ -57,7 +59,17 @@ export default function useSwapAdjustedAmounts(tradeDetails: Quote) {
     (tradeDetails.buyTokenAddress === ETH_ADDRESS &&
       tradeDetails.sellTokenAddress === WBNB_BSC_ADDRESS) ||
     (tradeDetails.sellTokenAddress === ETH_ADDRESS &&
-      tradeDetails.buyTokenAddress === WBNB_BSC_ADDRESS)
+      tradeDetails.buyTokenAddress === WBNB_BSC_ADDRESS) ||
+    // arb one <-> weth arb one swap
+    (tradeDetails.buyTokenAddress === ETH_ADDRESS &&
+      tradeDetails.sellTokenAddress === WETH_ARBITRUM_ADDRESS) ||
+    (tradeDetails.sellTokenAddress === ETH_ADDRESS &&
+      tradeDetails.buyTokenAddress === WETH_ARBITRUM_ADDRESS) ||
+    // zora eth <-> weth swap
+    (tradeDetails.buyTokenAddress === ETH_ADDRESS &&
+      tradeDetails.sellTokenAddress === WETH_ZORA_ADDRESS) ||
+    (tradeDetails.sellTokenAddress === ETH_ADDRESS &&
+      tradeDetails.buyTokenAddress === WETH_ZORA_ADDRESS)
   ) {
     amountReceivedSold = fromWei(amountReceivedSold.toString());
   }