+++ CHANGELOG.md
@@ -15,6 +15,57 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 
 ### Fixed
 
+## [1.9.10] (https://github.com/rainbow-me/rainbow/releases/tag/v1.9.10)
+
+### Added
+
+- Tx Simulation (#5177)
+- RPC Proxy updates (#5169)
+- Remote promo sheet capabilities (#5140)
+
+### Changed
+
+- ‘An error occurred’ popup changes (#5187)
+
+### Fixed
+
+- Android navigation bar now matches app theme (#5150)
+- Infinite render on swaps modal bug (#5191)
+
+## [1.9.9] (https://github.com/rainbow-me/rainbow/releases/tag/v1.9.9)
+
+### Added
+
+- Bug fixes
+- WC improvements
+
+## [1.9.8] (https://github.com/rainbow-me/rainbow/releases/tag/v1.9.8)
+
+### Added
+
+- WC dapp warnings
+- e2e updates
+- Fee updates to NFT Mints
+- Account Asset improvements
+
+## [1.9.7] (https://github.com/rainbow-me/rainbow/releases/tag/v1.9.7)
+
+### Added
+
+- points v0
+- prompt app reviews
+- bug fixes
+
+## [1.9.6] (https://github.com/rainbow-me/rainbow/releases/tag/v1.9.6)
+
+### Added
+
+- in app mints
+- i18n support
+- Rainbow Tabs
+- OP Rewards Round 2
+- more user pain points fixes
+
 ## [1.9.5] (https://github.com/rainbow-me/rainbow/releases/tag/v1.9.5)
 
 ### Added